Move Lazy signatures out of simpleapi into funcinspect
RE #28809
